Adding log capabilities to application.

+++ src/main/java/com/notifier/app/model/repository/NotifierUserRepository.java
@@ -9,7 +9,7 @@ public interface NotifierUserRepository extends JpaRepository<User, String> {
 
     List<User> findAll();
 
-    List<User> findAllByUserCategories_CategoryId(int id);
+    List<User> findAllByUserCategories_CategoryId(Long id);
 
-    List<User> findAllByUserChannels_ChannelId(int id);
+    List<User> findAllByUserChannels_ChannelId(Long id);
 }
